Move TestPort out of _unix
